+++ pysat/_orbits.py
@@ -1,11 +1,8 @@
-from __future__ import print_function
-from __future__ import absolute_import
-
 import datetime as dt
 import functools
-
 import numpy as np
 import pandas as pds
+
 from pysat import logger